refactor(parts): extract schemas to domain/parts/schemas.py (CQ-006)
Closes #122 (PR1 of the staged plan; prereq for #118's parts.py
split). Establishes the rule "every domain has its `schemas.py`;
routers do not declare inline `class XxxIn(BaseModel)`."

- backend/app/domain/parts/schemas.py — new module with PartIn,
  PartPatch, BulkDeleteIn, SubstituteIn, MetaMemberIn, ScanImportRow,
  ScanImportIn, QuickRemoveBagIn. Every class keeps
  `model_config = ConfigDict(extra="forbid")` (pinned by
  test_extra_forbid.py).
- backend/app/api/routes/parts.py — drops the eight inline classes
  and imports from the new module. Removes now-unused BaseModel /
  ConfigDict / Field / Literal imports. parts.py shrinks
  1245 → 1146 lines.
- docs/ARCHITECTURE.md — promotes the schema-location rule from a
  loose convention to an explicit invariant in the Domain
  decomposition section.

Pure organisational move — no runtime change. Out of scope: migrating
the other inlined routers (lots, tags, custom_fields, workspaces,
storage, etc.); that's PR2/PR3 of the plan.

v2 ID: CQ-006.
